Allow saving and forking of shaders

+++ shaderwall.py
@@ -1,6 +1,9 @@
 import bottle
 import sqlite3 as sql
 import base64
+import random
+import string
+import json
 
 bottle.BaseRequest.MEMFILE_MAX = 1024 * 1024 # 1MB
 
@@ -22,18 +25,28 @@ def get_gallery():
 @bottle.view('static/editor.html')
 def get_gallery(shader_id=None):
     if shader_id:
+        authcode = bottle.request.params.getunicode('authcode')
         cursor = conn.cursor()
-        cursor.execute('SELECT id, source FROM shader WHERE id = ?', (shader_id,))
+        cursor.execute('SELECT id, source, authcode FROM shader WHERE id = ?', (shader_id,))
         result = cursor.fetchone()
         if result:
+            if result[2] == authcode:
+                save_button_text = 'Save'
+                save_url = '/shaders/%d' % shader_id
+            else:
+                save_button_text = 'Fork'
+                save_url = '/shaders'
             return {
                 'shader_id': result[0],
                 'shader_source': result[1],
+                'save_button_text': save_button_text,
+                'save_url': save_url,
+                'authcode': authcode,
             }
         else:
-            return {}
+            return {'save_url': '/shaders', 'save_button_text': 'Create', 'authcode': ''}
     else:
-        return {}
+        return {'save_url': '/shaders', 'save_button_text': 'Create', 'authcode': ''}
 
 @bottle.route('/lib/<path:path>')
 def get_static(path):
@@ -43,25 +56,35 @@ def get_static(path):
 def get_screenshot(path):
     return bottle.static_file(path, root='./uploads')
 
+# generate 32byte authentication code
+def generate_authcode():
+    return ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(32))
+
+def save_screenshot(shader_id, screenshot):
+    try:
+        if not screenshot[:22] == 'data:image/png;base64,':
+            return False
+        screenshot = base64.b64decode(screenshot[22:])
+        open("uploads/%d.png" % shader_id, "w").write(screenshot);
+        return True
+    except:
+        return False
+
 @bottle.post('/shaders')
 def create_shader():
     source = bottle.request.params.getunicode('source')
     screenshot = bottle.request.params.getunicode('screenshot')
-    if not screenshot[:22] == 'data:image/png;base64,':
-        return bottle.abort(418, "I'm a teapot.")
 
     cursor = conn.cursor()
-    cursor.execute('INSERT INTO shader (source) VALUES (?)', (source,))
+    new_authcode = generate_authcode()
+    cursor.execute('INSERT INTO shader (source, authcode) VALUES (?,?)', (source,new_authcode))
     conn.commit()
     new_shader_id = cursor.lastrowid
-    try:
-        screenshot = base64.b64decode(screenshot[22:])
-        open("uploads/%d.png" % cursor.lastrowid, "w").write(screenshot);
-    except:
+    if not save_screenshot(new_shader_id, screenshot):
         cursor.execute('DELETE FROM shader WHERE id = ?', (new_shader_id))
         return bottle.abort(500, "Internal Server Error")
 
-    return 'Created new shader ' + str(cursor.lastrowid)
+    return json.dumps({'id': new_shader_id, 'authcode': new_authcode})
 
 @bottle.get('/shaders/<shader_id:int>')
 def get_shader(shader_id):
@@ -77,11 +100,15 @@ def get_shader(shader_id):
 def edit_shader(shader_id):
     source = bottle.request.params.getunicode('source')
     authcode = bottle.request.params.getunicode('authcode')
+    screenshot = bottle.request.params.getunicode('screenshot')
     cursor = conn.cursor()
     cursor.execute('UPDATE shader SET source = ? WHERE id = ? AND authcode = ?' , (source, shader_id, authcode))
     conn.commit()
+    if not save_screenshot(shader_id, screenshot):
+        print("yo?")
+
     if cursor.rowcount == 1:
-        return 'Updated shader'
+        return json.dumps({'id': shader_id, 'authcode': authcode})
     else:
         bottle.abort(403, 'Update failed')
